+++ example/lib/native_app.dart
@@ -16,7 +16,6 @@ import 'pages/header_page.dart';
 import 'pages/icon_page.dart';
 import 'pages/input_trigger_page.dart';
 import 'pages/link_page.dart';
-import 'pages/list_header_page.dart';
 import 'pages/list_item_page.dart';
 import 'pages/loading_indicator_page.dart';
 import 'pages/message_page.dart';
@@ -110,7 +109,6 @@ class MyApp extends StatelessWidget {
                                   _DemoEntry('Checkbox', CheckboxPage()),
                                   _DemoEntry('Chip', ChipPage()),
                                   _DemoEntry('Link', LinkPage()),
-                                  _DemoEntry('List Header', ListHeaderPage()),
                                   _DemoEntry('List Item', ListItemPage()),
                                   _DemoEntry('Loading Indicator', LoadingIndicatorPage()),
                                   _DemoEntry('Picker', PickerPage()),
@@ -138,7 +136,7 @@ class MyApp extends StatelessWidget {
                                 context: context,
                                 items: [
                                   _DemoEntry('Autocompletion', AutocompletionPage()),
-                                  _DemoEntry('Group', ContainerPage()),
+                                  _DemoEntry('Container', ContainerPage()),
                                   _DemoEntry('Header', HeaderPage()),
                                   _DemoEntry('Headerbox', HeaderBoxPage()),
                                   _DemoEntry('Modal', ModalPage()),